requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn>=20.1.0

sqlalchemy
requests>=2.28.0
python-multipart
pydantic
pydantic-settings
gffutils
pyensembl
pyliftover>=0.4
pandas>=1.5.0
scikit-learn>=1.3.0
torch>=2.0.0
biopython>=1.81
openpyxl>=3.1.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

gradio
sentence-transformers

pyliftover==0.4
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.2.0
numpy>=1.24.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3
hdbscan>=0.8.32
stable-baselines3>=2.1.0
gym>=0.21.0
